update grade exam

+++ BE/GESS.Model/QuestionPracExamDTO/QuestionPracExamDTO.cs
@@ -11,7 +11,7 @@ namespace GESS.Model.QuestionPracExam
         public Guid PracExamHistoryId { get; set; }
         public string QuestionContent { get; set; }
         public string? Answer { get; set; }
-        public int ? Score { get; set; }
+        public double ? Score { get; set; }
         public string? GradingCriteria { get; set; }
     }
 }
